cfintraday/requirements.txt
google-cloud-storage==1.42.0
google-cloud-scheduler==2.3.2
google-api-python-client==2.17.0
grpcio==1.39.0